backend/requirements.txt
# Core web and agent frameworks
fastapi
google-adk==1.14.0
a2a-sdk>=0.3.0

# Google Cloud SDK for Vertex AI
google-cloud-aiplatform

# Database connector
psycopg2-binary
google-cloud-alloydb-connector[pg8000,psycopg2]

# SQLAlchemy ORM
sqlalchemy>=2.0.0
alembic>=1.13.0
pgvector>=0.3.0

# Configuration management
pydantic
pydantic-settings
python-dotenv

# ASGI Server
uvicorn

# HTTP client
requests

# Google Cloud Secret Manager
google-cloud-secret-manager
